malloc: fix malloc of size 0

+++ src/tools.h
@@ -11,7 +11,7 @@
 #define IS_SET(NUM, POS) ((NUM) & (1 << (POS)))
 
 int mark_block(size_t *free_list, size_t size);
-void *get_block(void *bucket, size_t n, size_t block_size);
+void *get_block(void *bucket, int n, size_t block_size);
 void *page_begin(void *ptr, size_t page_size);
 void set_free(size_t pos, size_t *free_list);
 void reset_list(size_t *free_list, size_t block_size);
